+++ src/yeezy/app.py
@@ -495,79 +495,3 @@ class Yeezy:
         for func_name, properties in self.functions.items():
             self.log(f"Function: {func_name}, properties: {properties}")
         self.log("-" * 100)
-
-
-if __name__ == "__main__":
-    import torch
-
-    yee = Yeezy(__name__, debug=True, log_path=None)
-    item = Troll()
-
-
-    # This should register all functions inside of class Test()
-    # @yee.time
-    class Test:
-        def __init__(self):
-            self.test = [1, 2, 3]
-
-        def mutate(self, num):
-            self.test.append(num)
-
-        def print_something(self, test):
-            print(test)
-
-        # @yee.trace()
-        @yee.time
-        def create_long_list(self, n=1000000, name="test"):
-            return torch.tensor(range(n))
-
-
-    def print_something(name):
-        print(name)
-        num = 10
-        # yee.debug(num)
-
-
-    def do_go(num: list):
-        print(f"yee registered running")
-
-
-    @yee.time
-    def create_long_list(n=1000000, name="test"):
-        return list(range(n)), name
-
-
-    # create_long_list = yee.double_wrap(create_long_list)
-    tom = Test()
-
-
-    def deco(func):
-        @wraps(func)
-        def yee(*args, **kwargs):
-            print("yeet")
-            return func(*args, **kwargs)
-        return yee
-
-   # create_long_list = deco(create_long_list)
-
-    # We can decorate methods using the following methods
-    troll = Troll()
-    # another_fn = yee.decorate_before(troll.print_this)
-    # Another create long list
-    def bonk(*args, **kwargs):
-        print(f"Bonk: {args}, {kwargs}")
-
-    fn = yee.before(cll)
-    fn = yee.before(bonk)
-
-    for i in range(20):
-        tom.create_long_list(i)
-        fn(i)
-        # create_long_list(100)
-        # tom.create_long_list(1000)
-        # tom.mutate(10)
-
-    print("-" * 100)
-
-    # Profile and gather information
-    yee.analyze()
