+++ src/yeezy/demo.py
@@ -1,0 +1,52 @@
+from app import Yeezy
+
+
+def simple_demo():
+    import torch
+
+    yee = Yeezy(__name__, debug=True, log_path=None)
+
+    # This should register all functions inside of class Test()
+    # @yee.time
+    class Test:
+        def __init__(self):
+            self.test = [1, 2, 3]
+
+        def mutate(self, num):
+            self.test.append(num)
+
+        def print_something(self, test):
+            print(test)
+
+        # @yee.trace()
+        @yee.time
+        def create_long_list(self, n=1000000, name="test"):
+            return torch.tensor(range(n))
+
+
+    @yee.time
+    def create_long_list(n=1000000, name="test"):
+        return list(range(n)), name
+
+    # Some class
+    tom = Test()
+
+    # Another create long list
+    def bonk(*args, **kwargs):
+        print(f"Bonk: {args}, {kwargs}")
+
+    # We can also add decorators dynamically
+    create_long_list = yee.before(bonk)
+
+    for i in range(20):
+        tom.create_long_list(i)
+        create_long_list(i)
+
+    print("-" * 100)
+
+    # Profile and gather information
+    yee.analyze()
+
+
+if __name__ == "__main__":
+    simple_demo()
